requirements.txt
Django==4.2.11
python-dotenv==1.0.1
twilio==9.3.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-caching-md5==1.0.2
django-debug-toolbar==4.2.0
drf-spectacular==0.26.5
cryptography==41.0.7
django-push-notifications==3.0.2
django-apscheduler==0.6.2
google-cloud-storage==2.10.0
boto3==1.28.0
pydub==0.25.1
python-magic==0.4.27
channels==4.0.0
channels-redis==4.1.0
django-model-utils==4.3.1
factory-boy==3.3.0
coverage==7.3.0
black==23.11.0
flake8==6.1.0
packaging==23.2
Sentry-SDK==1.38.0
django-ratelimit==4.1.0
graphene-django==3.1.1
python-dateutil==2.8.2
